Bugfix add control user id

+++ Application/Features/User/Commands/Delete/DeleteUserCommand.cs
@@ -1,10 +1,13 @@
 ﻿using Application.Repositories;
 using AutoMapper;
 using Core.Application.Pipelines.Authorization;
+using Core.CrossCuttingConcerns.Exceptions.Types;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -19,17 +22,23 @@ namespace Application.Features.User.Commands.Delete
 
             private readonly IUserRepository _userRepository;
             private readonly IMapper _mapper;
+            private readonly IHttpContextAccessor _httpContextAccessor;
 
-            public UserDeleteCommandHandler(IUserRepository userRepository, IMapper mapper)
+            public UserDeleteCommandHandler(IUserRepository userRepository, IMapper mapper, IHttpContextAccessor httpContextAccessor)
             {
                 _userRepository = userRepository;
                 _mapper = mapper;
+                _httpContextAccessor = httpContextAccessor;
             }
             public async Task<DeleteUserCommandResponse> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
             {
-                var userToDelete = await _userRepository.GetAsync(u => u.Id == request.UserId);
-                await _userRepository.DeleteAsync(userToDelete);
-                return new() { };
+                var userId = int.Parse(_httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
+                if (userId == request.UserId)
+                {
+                    var userToDelete = await _userRepository.GetAsync(u => u.Id == request.UserId);
+                    await _userRepository.DeleteAsync(userToDelete);
+                }
+                throw new BusinessException("İşlem başarısız");
             }
         }
     }
